requirements.txt
absl-py
astunparse
cached-property
cachetools
certifi
chardet
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
Keras
Keras-Preprocessing
Markdown
matplotlib
numpy
oauthlib
opencv-python
opt-einsum
Pillow
protobuf
pyasn1
pyasn1-modules
PyYAML
requests
requests-oauthlib
rsa
scipy
six
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
typing-extensions
urllib3
Werkzeug
wrapt
zipp
scikit-learn
